pyproject.toml
[project]
name = "strong-hyperliquid-websocket"
version = "0.1.0"
description = "Hyperliquid WebSocket API 增强连接管理工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hyperliquid-python-sdk>=0.21.0",
    "aiohttp>=3.9.0",
    "aiolimiter>=1.1.0",
    "asyncpg>=0.29.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
